Refactor import statements to use relative imports consistently
